+++ main.go
@@ -17,13 +17,13 @@ import (
 
 	_ "github.com/bwNetFlow/flowpipeline/segments/alert/http"
 
-	_ "github.com/bwNetFlow/flowpipeline/segments/controlflow/blackhole"
 	_ "github.com/bwNetFlow/flowpipeline/segments/controlflow/skip"
 	_ "github.com/bwNetFlow/flowpipeline/segments/controlflow/subpipeline"
 
 	_ "github.com/bwNetFlow/flowpipeline/segments/export/influx"
 	_ "github.com/bwNetFlow/flowpipeline/segments/export/prometheus"
 
+	_ "github.com/bwNetFlow/flowpipeline/segments/filter/drop"
 	_ "github.com/bwNetFlow/flowpipeline/segments/filter/elephant"
 	_ "github.com/bwNetFlow/flowpipeline/segments/filter/flowfilter"
